+++ packages/celest/test/runtime/data/connect_test.dart
@@ -2,16 +2,25 @@
 @Tags(['e2e'])
 library;
 
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io' as io;
 import 'dart:io';
+import 'dart:isolate';
 
 import 'package:celest/src/config/config_values.dart';
 import 'package:celest/src/core/context.dart';
 import 'package:celest/src/runtime/data/connect.dart';
-import 'package:drift/drift.dart';
+import 'package:drift/native.dart';
 import 'package:drift_hrana/drift_hrana.dart';
+import 'package:path/path.dart' as p;
 import 'package:platform/platform.dart';
-import 'package:test/fake.dart';
+import 'package:process/process.dart';
 import 'package:test/test.dart';
+
+import 'test_database.dart';
+
+const processManager = LocalProcessManager();
 
 Future<int> _findOpenPort() async {
   final server = await ServerSocket.bind(InternetAddress.loopbackIPv4, 0);
@@ -20,52 +29,119 @@ Future<int> _findOpenPort() async {
   return port;
 }
 
-Future<Uri> _startLibSqlServer() async {
+Future<Uri> _startSqld() async {
   final port = await _findOpenPort();
-  final process = await Process.start('turso', ['dev', '--port', '$port']);
+  final process = await processManager.start([
+    if (io.Platform.environment.containsKey('CI'))
+      '/home/runner/.turso/turso'
+    else
+      'turso',
+    'dev',
+    '--port',
+    '$port',
+  ]);
   addTearDown(process.kill);
+  final running = Completer<void>();
+  process.stdout
+      .transform(utf8.decoder)
+      .transform(const LineSplitter())
+      .listen((event) {
+    if (event.contains('sqld listening on')) {
+      running.complete();
+    }
+  });
+  await running.future.timeout(
+    const Duration(seconds: 5),
+    onTimeout: () {
+      throw StateError('Failed to start sqld');
+    },
+  );
   return Uri.parse('ws://localhost:$port');
 }
 
 void main() {
   group('connect', () {
-    test('allows local libSQL URIs', () async {
-      final host = await _startLibSqlServer();
-      final platform = FakePlatform(
-        environment: {
-          'CELEST_DATABASE_HOST': host.toString(),
-        },
-      );
-      Context.current.put(env.environment, 'production');
+    test(
+      'allows local sqld URIs',
+      // TODO(dnys1): Get Turso CLI working in CI
+      skip: io.Platform.environment.containsKey('CI'),
+      () async {
+        final host = await _startSqld();
+        final platform = FakePlatform(
+          environment: {
+            'CELEST_DATABASE_HOST': host.toString(),
+          },
+        );
+        Context.current.put(env.environment, 'production');
+        Context.current.put(ContextKey.platform, platform);
+        final database = await connect(
+          Context.current,
+          name: 'test',
+          factory: expectAsync1((executor) {
+            expect(executor, isA<HranaDatabase>());
+            return TestDatabase(executor);
+          }),
+          hostnameVariable: const env('CELEST_DATABASE_HOST'),
+          tokenSecret: const secret('CELEST_DATABASE_TOKEN'),
+        );
+        await database.close();
+      },
+    );
+  });
+
+  group('localExecutor', () {
+    test('uses package config when path=null', () async {
+      final packageConfig = await Isolate.packageConfig;
+      final file = File.fromUri(packageConfig!.resolve('./celest/test.db'));
+      if (file.existsSync()) {
+        file.deleteSync();
+      }
+      addTearDown(() {
+        if (file.existsSync()) {
+          file.deleteSync();
+        }
+      });
+
+      final platform = FakePlatform();
+      Context.current.put(env.environment, 'local');
       Context.current.put(ContextKey.platform, platform);
-      await connect(
+      final database = await connect(
         Context.current,
         name: 'test',
         factory: expectAsync1((executor) {
-          expect(executor, isA<HranaDatabase>());
-          return _FakeDatabase();
+          expect(executor, isA<NativeDatabase>());
+          return TestDatabase(executor);
         }),
         hostnameVariable: const env('CELEST_DATABASE_HOST'),
         tokenSecret: const secret('CELEST_DATABASE_TOKEN'),
       );
+      addTearDown(database.close);
+
+      expect(file.existsSync(), isTrue);
+    });
+
+    test('path != null', () async {
+      final tmpDir = await Directory.systemTemp.createTemp('celest_test');
+      addTearDown(() => tmpDir.delete(recursive: true));
+
+      final platform = FakePlatform();
+      Context.current.put(env.environment, 'local');
+      Context.current.put(ContextKey.platform, platform);
+      final database = await connect(
+        Context.current,
+        name: 'test',
+        factory: expectAsync1((executor) {
+          expect(executor, isA<NativeDatabase>());
+          return TestDatabase(executor);
+        }),
+        hostnameVariable: const env('CELEST_DATABASE_HOST'),
+        tokenSecret: const secret('CELEST_DATABASE_TOKEN'),
+        path: p.join(tmpDir.path, 'test.db'),
+      );
+      addTearDown(database.close);
+
+      final file = File.fromUri(tmpDir.uri.resolve('./test.db'));
+      expect(file.existsSync(), isTrue);
     });
   });
-}
-
-final class _FakeDatabase extends Fake implements GeneratedDatabase {
-  @override
-  Selectable<QueryRow> customSelect(
-    String query, {
-    List<Variable<Object>> variables = const [],
-    Set<ResultSetImplementation<dynamic, dynamic>> readsFrom = const {},
-  }) {
-    return _FakeSelectable();
-  }
-}
-
-final class _FakeSelectable extends Fake implements Selectable<Never> {
-  @override
-  Future<List<Never>> get() {
-    return Future.value([]);
-  }
 }