+++ packages/celest/test/runtime/data/test_database.dart
@@ -1,0 +1,16 @@
+import 'package:drift/drift.dart';
+
+part 'test_database.g.dart';
+
+class Items extends Table {
+  IntColumn get id => integer().autoIncrement()();
+  TextColumn get name => text().nullable()();
+}
+
+@DriftDatabase(tables: [Items])
+class TestDatabase extends _$TestDatabase {
+  TestDatabase(super.e);
+
+  @override
+  int get schemaVersion => 1;
+}